+++ components/api-server/src/schema/eventsMethods.js
@@ -24,7 +24,7 @@ const boolean = helpers.boolean;
 module.exports = {
   get: {
     params: object({
-      'streams': array(string()),
+      'streams': {},
       'tags': array(string()),
       'types': array(string()),
       'fromTime': number(),